@@ -6,26 +6,28 @@
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "lint": "ts-standard",
+    "lint:fix": "ts-standard --fix",
     "preview": "vite preview",
     "test": "jest"
   },
   "dependencies": {
     "@heroicons/react": "1.0.6",
-    "@tremor/react": "^3.13.1",
     "axios": "^1.6.5",
     "i18next": "^23.7.16",
+    "jwt-decode": "^4.0.0",
     "localforage": "^1.10.0",
     "match-sorter": "^6.3.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-i18next": "14.0.0",
+    "react-modal": "^3.16.1",
     "react-router-dom": "^6.21.1",
+    "react-toastify": "^10.0.4",
+    "react-tooltip": "^5.25.2",
     "sort-by": "^1.2.0",
+    "sweetalert2": "^11.10.4",
     "zustand": "4.4.7"
-  },
-  "eslintConfig": {
-    "extends": "./node_modules/standard/eslintrc.json"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.2.0",
@@ -33,21 +35,22 @@
     "@types/jest": "^29.5.11",
     "@types/react": "^18.2.47",
     "@types/react-dom": "^18.2.18",
-    "@typescript-eslint/eslint-plugin": "^6.14.0",
-    "@typescript-eslint/parser": "^6.14.0",
+    "@typescript-eslint/eslint-plugin": "^6.18.1",
+    "@typescript-eslint/parser": "^6.18.1",
     "@vitejs/plugin-react-swc": "^3.5.0",
     "autoprefixer": "^10.4.16",
     "babel-jest": "^29.7.0",
-    "eslint": "^8.55.0",
+    "eslint": "^8.56.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.5",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "postcss": "^8.4.33",
+    "standard": "^17.1.0",
     "tailwindcss": "^3.4.1",
     "ts-jest": "^29.1.1",
     "ts-standard": "^12.0.2",
-    "typescript": "^5.2.2",
+    "typescript": "^5.3.3",
     "vite": "^5.0.8"
   }
 }
