@@ -1,12 +1,17 @@
-import React from "react";
+import React from 'react'
 
 interface Props {
-  children: React.ReactNode;
+  children: React.ReactNode
+  className?: string
+  id?: string
 }
-export const Card = ({ children }: Props) => {
+export const Card = ({ children, className, id }: Props) => {
   return (
-    <div className="w-full h-full rounded bg-white/90 dark:bg-gray-500/90 p-5">
+    <div
+      id={id}
+      className={`relative overflow-x-auto w-full rounded bg-white/90 dark:bg-gray-900 dark:text-white  p-2 lg:p-5   ${className}`}
+    >
       {children}
     </div>
-  );
-};
+  )
+}
