@@ -1,8 +1,9 @@
-import { NavBar } from "./Navigation/NavBar"
+import { NavBar } from './Navigation/NavBar'
 
-
-export const Header = ()=>{
-    return (<header className={`[grid-area:header]`}>
-        <NavBar />
-    </header>)
+export const Header = () => {
+  return (
+    <header className='[grid-area:header]'>
+      <NavBar />
+    </header>
+  )
 }
