@@ -1,42 +1,42 @@
 
-export const BarIcon = (props:any) => {
-  const { size=6 } = props;
+export const BarIcon = (props: any) => {
+  const { size = 6 } = props
   return (
     <svg
-      xmlns="http://www.w3.org/2000/svg"
-      fill="none"
-      viewBox="0 0 24 24"
-      strokeWidth="1.5"
-      stroke="currentColor"
-      data-slot="icon"
+      xmlns='http://www.w3.org/2000/svg'
+      fill='none'
+      viewBox='0 0 24 24'
+      strokeWidth='1.5'
+      stroke='currentColor'
+      data-slot='icon'
       className={`w-${size} h-${size} `}
     >
       <path
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        d="M3.75 6.75h16.5M3.75 12h16.5m-16.5 5.25h16.5"
+        strokeLinecap='round'
+        strokeLinejoin='round'
+        d='M3.75 6.75h16.5M3.75 12h16.5m-16.5 5.25h16.5'
       />
     </svg>
-  );
+  )
 }
 
-export const BarIconBottomRight = (props:any) => {
-  const { size=6 } = props;
+export const BarIconBottomRight = (props: any) => {
+  const { size = 6 } = props
   return (
     <svg
-      xmlns="http://www.w3.org/2000/svg"
-      fill="none"
-      viewBox="0 0 24 24"
-      strokeWidth="1.5"
-      stroke="currentColor"
-      data-slot="icon"
+      xmlns='http://www.w3.org/2000/svg'
+      fill='none'
+      viewBox='0 0 24 24'
+      strokeWidth='1.5'
+      stroke='currentColor'
+      data-slot='icon'
       className={`w-${size} h-${size}`}
     >
       <path
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        d="M3.75 6.75h16.5M3.75 12h16.5M12 17.25h8.25"
+        strokeLinecap='round'
+        strokeLinejoin='round'
+        d='M3.75 6.75h16.5M3.75 12h16.5M12 17.25h8.25'
       />
     </svg>
-  );
+  )
 }
