@@ -1,20 +1,63 @@
-
-export const ChevronDown = (props:any) => {
-  const { size=6 } = props;
+interface Props {
+  size?: number
+  className?: string
+}
+export const ChevronDown = (props: Props) => {
+  const { size = 6, className } = props
   return (
     <svg
-      xmlns="http://www.w3.org/2000/svg"
-      fill="none"
-      viewBox="0 0 24 24"
-      strokeWidth="1.5"
-      stroke="currentColor"
-      className={`w-${size} h-${size}`}
+      xmlns='http://www.w3.org/2000/svg'
+      fill='none'
+      viewBox='0 0 24 24'
+      strokeWidth='1.5'
+      stroke='currentColor'
+      className={`size-${size} ${className}`}
     >
       <path
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        d="m19.5 8.25-7.5 7.5-7.5-7.5"
+        strokeLinecap='round'
+        strokeLinejoin='round'
+        d='m19.5 8.25-7.5 7.5-7.5-7.5'
       />
     </svg>
-  );
-};
+  )
+}
+
+export const ChevronLeft = (props: Props) => {
+  const { size = 6, className } = props
+  return (
+    <svg
+      xmlns='http://www.w3.org/2000/svg'
+      fill='none'
+      viewBox='0 0 24 24'
+      strokeWidth={1.5}
+      stroke='currentColor'
+      className={`size-${size} ${className}`}
+    >
+      <path
+        strokeLinecap='round'
+        strokeLinejoin='round'
+        d='M15.75 19.5 8.25 12l7.5-7.5'
+      />
+    </svg>
+  )
+}
+
+export const ChevronRight = (props: Props) => {
+  const { size = 6, className } = props
+  return (
+    <svg
+      xmlns='http://www.w3.org/2000/svg'
+      fill='none'
+      viewBox='0 0 24 24'
+      strokeWidth={1.5}
+      stroke='currentColor'
+      className={`size-${size} ${className}`}
+    >
+      <path
+        strokeLinecap='round'
+        strokeLinejoin='round'
+        d='m8.25 4.5 7.5 7.5-7.5 7.5'
+      />
+    </svg>
+  )
+}
