@@ -1,19 +1,19 @@
 export const LogOutIcon = () => {
   return (
     <svg
-      fill="none"
-      height="24"
-      shapeRendering="geometricPrecision"
-      stroke="currentColor"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth="1.5"
-      viewBox="0 0 24 24"
-      width="24"
+      fill='none'
+      height='24'
+      shapeRendering='geometricPrecision'
+      stroke='currentColor'
+      strokeLinecap='round'
+      strokeLinejoin='round'
+      strokeWidth='1.5'
+      viewBox='0 0 24 24'
+      width='24'
     >
-      <path d="M9 21H5a2 2 0 01-2-2V5a2 2 0 012-2h4" />
-      <path d="M16 17l5-5-5-5" />
-      <path d="M21 12H9" />
+      <path d='M9 21H5a2 2 0 01-2-2V5a2 2 0 012-2h4' />
+      <path d='M16 17l5-5-5-5' />
+      <path d='M21 12H9' />
     </svg>
-  );
-};
+  )
+}
