@@ -1,20 +1,23 @@
-export const MoonIcon = (props: any) => {
-  const { size = 6, css = "" } = props;
+interface Props {
+  size: number
+  css?: string
+}
+export const MoonIcon = ({ size = 6, css = '' }: Props) => {
   return (
     <svg
-      xmlns="http://www.w3.org/2000/svg"
-      fill="none"
-      viewBox="0 0 24 24"
-      strokeWidth="1.5"
-      stroke="currentColor"
-      data-slot="icon"
-      className={`w-${size} h-${size} ${css} `}
+      xmlns='http://www.w3.org/2000/svg'
+      fill='none'
+      viewBox='0 0 24 24'
+      strokeWidth='1.5'
+      stroke='currentColor'
+      data-slot='icon'
+      className={`size-${size} ${css} `}
     >
       <path
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        d="M21.752 15.002A9.72 9.72 0 0 1 18 15.75c-5.385 0-9.75-4.365-9.75-9.75 0-1.33.266-2.597.748-3.752A9.753 9.753 0 0 0 3 11.25C3 16.635 7.365 21 12.75 21a9.753 9.753 0 0 0 9.002-5.998Z"
+        strokeLinecap='round'
+        strokeLinejoin='round'
+        d='M21.752 15.002A9.72 9.72 0 0 1 18 15.75c-5.385 0-9.75-4.365-9.75-9.75 0-1.33.266-2.597.748-3.752A9.753 9.753 0 0 0 3 11.25C3 16.635 7.365 21 12.75 21a9.753 9.753 0 0 0 9.002-5.998Z'
       />
     </svg>
-  );
-};
+  )
+}
