@@ -1,41 +1,42 @@
-import { useTranslation } from "react-i18next";
-import { AuthorizeRutes } from "../../routes/AuthorizeRutes";
-import { OnlyAuthenticateRotes } from "../../routes/OnlyAuthenticateRoutes";
-import { useSideNavStores } from "../../stores/ui/sidenavbar";
-import { LogOutIcon } from "../Icons/LogOutIcon";
-import SideNavBarCollapseOption from "./SideNavBarCollapseOption";
-import { SideNavBarOption, SideNavBarOptionCallback } from "./SideNavBarOption";
-import { Divider } from "@tremor/react";
+import { useTranslation } from 'react-i18next'
+import { AuthorizeRutes } from '../../routes/AuthorizeRutes'
+import { OnlyAuthenticateRotes } from '../../routes/OnlyAuthenticateRoutes'
+import { useSideNavStores } from '../../stores/ui/sidenavbar'
+import { LogOutIcon } from '../Icons/LogOutIcon'
+import SideNavBarCollapseOption from './SideNavBarCollapseOption'
+import { SideNavBarOption, SideNavBarOptionCallback } from './SideNavBarOption'
+import { useLoginStore } from '../../stores/auth'
 
 export const SideNavBar = () => {
-  const mini = useSideNavStores((state) => state.mini);
-  const { t } = useTranslation();
-  const SideNavRoutes = AuthorizeRutes;
+  const mini = useSideNavStores((state) => state.mini)
+  const logOut = useLoginStore((state) => state.logOut)
+  const { t } = useTranslation()
+  const SideNavRoutes = AuthorizeRutes
 
   return (
     <aside
       className={`[grid-area:aside] ${
-        mini ? "aside-mini" : "aside"
-      } transition-all delay-75 ease-out `}
+        mini ? 'aside-mini' : 'aside'
+      } transition-all delay-75 ease-out z-10 `}
     >
-      <div className="aside-nav-wrapper">
-        <div className="aside-nav-head">
+      <div className='aside-nav-wrapper'>
+        <div className='aside-nav-head'>
           <picture>
             <img
-              src="https://placehold.co/50x50?text=Logo"
-              alt="Logo"
-              className="rounded-full"
+              src='https://placehold.co/50x50?text=Logo'
+              alt='Logo'
+              className='rounded-full'
             />
           </picture>
-          {!mini && <h1 className="lg:block">Dashboard</h1>}
+          {!mini && <h1 className='lg:block'>Dashboard</h1>}
         </div>
 
-        <hr className="mt-2" />
+        <hr className='mt-2' />
 
-        <div className="py-2 h-[92%] flex flex-col justify-between">
+        <div className='py-2 min-h-dvh flex flex-col justify-between'>
           <ul>
             {SideNavRoutes.map((route, index) => {
-              const children = route?.children;
+              const children = route?.children
               if (children?.length) {
                 return (
                   <SideNavBarCollapseOption
@@ -54,30 +55,28 @@ export const SideNavBar = () => {
                       />
                     ))}
                   </SideNavBarCollapseOption>
-                );
+                )
               } else {
                 return (
                   <SideNavBarOption
                     key={index}
                     icon={route.icon}
                     label={route.label}
-                    path={route.path || "/"}
+                    path={route.path || '/'}
                     isMini={mini}
                   />
-                );
+                )
               }
             })}
           </ul>
 
           <ul>
-            <Divider />
+            <hr className='mt-2' />
             <SideNavBarOptionCallback
-              key={""}
+              key='logout'
               icon={<LogOutIcon />}
-              label={t("sideNavBar.logout")}
-              callback={() => {
-                alert("hi");
-              }}
+              label={t('sideNavBar.logout')}
+              callback={logOut}
               isMini={mini}
             />
 
@@ -94,5 +93,5 @@ export const SideNavBar = () => {
         </div>
       </div>
     </aside>
-  );
-};
+  )
+}
