@@ -1,3 +1,3 @@
-export const HomePage =  () =>{
-    return <h1>Home Page</h1>
+export const HomePage = () => {
+  return <h1>Home Page</h1>
 }
