@@ -1,20 +1,20 @@
-import { Card, Text } from "@tremor/react";
-import { NavSettings } from "../components/Settings/NavSettings";
-import { SettingLanguage } from "../components/Settings/Language";
-import { SettingAccount } from "../components/Settings/Account";
+import { Card } from '../components/Card'
+import { SettingAccount } from '../components/Settings/Account'
+import { SettingLanguage } from '../components/Settings/Language'
+import { NavSettings } from '../components/Settings/NavSettings'
 export const SettingsPage = () => {
   return (
     <>
-      <div className="flex gap-2">
-        <Card className="w-1/4">
-          <Text>Settings</Text>
+      <div className='flex flex-row gap-2'>
+        <Card className='w-1/5'>
+          <h1>Settings</h1>
           <NavSettings />
         </Card>
-        <Card className="grow flex flex-col gap-y-4">
+        <Card className='grow flex flex-col gap-y-4'>
           <SettingAccount />
           <SettingLanguage />
         </Card>
       </div>
     </>
-  );
-};
+  )
+}
