@@ -1,12 +1,12 @@
-import { Routes } from "../routes/Routes";
-import "./App.css";
+import { Routes } from '../routes/Routes'
+import './App.css'
 
-function App() {
+function App () {
   return (
     <>
       <Routes />
     </>
-  );
+  )
 }
 
-export default App;
+export default App
