@@ -1,8 +1,9 @@
-import { create } from "zustand";
-import { persist } from "zustand/middleware";
-import { IAuthState, ISession } from "../../abstraction/Interfazes/IAuth";
-import { LogInService } from "../../services/Auth";
-import { PREFIX_STORAGE } from "../../setup/constants";
+import axios from 'axios'
+import { create } from 'zustand'
+import { persist } from 'zustand/middleware'
+import { IAuthState, ISession } from '../../abstraction/Interfaces/IAuth'
+import { LogInService } from '../../services/Auth'
+import { PREFIX_STORAGE } from '../../setup/constants'
 
 export const useLoginStore = create<IAuthState>()(
   persist(
@@ -12,30 +13,33 @@ export const useLoginStore = create<IAuthState>()(
         errors: [],
         logIn: async ({
           username,
-          password,
+          password
         }: {
-          username: string;
-          password: string;
+          username: string
+          password: string
         }) => {
-          const result = await LogInService({ username, password });
+          const result = await LogInService({ username, password })
+
           const session: ISession = {
             token: result.token,
             user: result.user,
-            routes: result.routes,
-          };
+            routes: result.routes
+          }
 
           if (result.isSuccess) {
-            set({ session });
+            axios.defaults.headers.common.Authorization = `Bearer ${result.token}`
+            set({ session })
           } else {
-            set({ errors: result.errors });
+            set({ errors: result.errors })
           }
         },
         logOut: () => {
-          const session: ISession = { token: "" };
-          set({ session, errors: [] });
-        },
-      };
+          const session: ISession = { token: '' }
+          axios.defaults.headers.common.Authorization = null
+          set({ session, errors: [] })
+        }
+      }
     },
     { name: `${PREFIX_STORAGE}::SESSION` }
   )
-);
+)
