@@ -1,10 +1,10 @@
 export const Languages = [
   {
-    code: "en",
-    name: "english",
+    code: 'en',
+    name: 'english'
   },
   {
-    code: "es",
-    name: "spanish",
-  },
-];
+    code: 'es',
+    name: 'spanish'
+  }
+]
